requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
blinker==1.9.0
Brotli==1.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
ConfigArgParse==1.7.1
cryptography==45.0.4
Django==5.1.6
django-ninja==1.3.0
django-ninja-simple-jwt==0.6.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
Flask==3.1.1
flask-cors==6.0.1
Flask-Login==0.6.3
gevent==25.5.1
geventhttpclient==2.3.4
greenlet==3.2.3
idna==3.10
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
locust==2.32.10
MarkupSafe==3.0.2
msgpack==1.1.1
packaging==25.0
pillow==11.1.0
psutil==7.0.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.9.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
requests==2.32.4
setuptools==80.9.0
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
zope.event==5.0
zope.interface==7.2
